git-svn-id: svn://svn.forge.objectweb.org/svnroot/spagobi@1167 99afaf0d-6903-0410-885a-c66a8bbb5f81

+++ branch/SpagoBI-progress-base1.8/QbeCore/src/it/eng/qbe/services/ReportServlet.java
@@ -153,9 +153,9 @@ public class ReportServlet extends HttpServlet{
 			copyErrorMessageToResponse(response, e);
 			
 		}finally{
-			if (templateFile.exists())
+			if (templateFile != null && templateFile.exists())
 				templateFile.delete();
-			if (reportFile.exists())
+			if (reportFile != null && reportFile.exists())
 				reportFile.delete();
 			if ((aSession != null) && (aSession.isOpen())) 
 				aSession.close();
